Fixed conflicts after rebase on master
