+++ systemtest/src/main/java/io/strimzi/systemtest/utils/LogCollector.java
@@ -5,7 +5,7 @@
 package io.strimzi.systemtest.utils;
 
 import io.strimzi.systemtest.AbstractST;
-import io.strimzi.test.k8s.Kubernetes;
+import io.strimzi.test.k8s.KubeClient;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -14,16 +14,16 @@ import java.io.File;
 import static io.strimzi.test.BaseITST.cmdKubeClient;
 import static io.strimzi.test.TestUtils.writeFile;
 
-public class LogCollector {
+class LogCollector {
     private static final Logger LOGGER = LogManager.getLogger(AbstractST.class);
 
-    Kubernetes kubeClient;
-    String namespace;
-    File logDir;
-    File configMapDir;
-    File eventsDir;
+    private KubeClient kubeClient;
+    private String namespace;
+    private File logDir;
+    private File configMapDir;
+    private File eventsDir;
 
-    public LogCollector(Kubernetes kubeClient, File logDir) {
+    LogCollector(KubeClient kubeClient, File logDir) {
         this.kubeClient = kubeClient;
         this.namespace = kubeClient.getNamespace();
         this.logDir = logDir;
@@ -39,7 +39,7 @@ public class LogCollector {
         }
     }
 
-    public void collectLogsFromPods() {
+    void collectLogsFromPods() {
         LOGGER.info("Collecting logs for pods in namespace {}", namespace);
 
         try {
@@ -59,14 +59,14 @@ public class LogCollector {
         }
     }
 
-    public void collectEvents() {
+    void collectEvents() {
         LOGGER.info("Collecting events in namespace {}", namespace);
         String events = cmdKubeClient().getEvents();
         // Write events to file
         writeFile(eventsDir + "/" + "events-in-namespace" + kubeClient.getNamespace() + ".log", events);
     }
 
-    public void collectConfigMaps() {
+    void collectConfigMaps() {
         LOGGER.info("Collecting configmaps in namespace {}", namespace);
         kubeClient.listConfigMaps().forEach(configMap -> {
             writeFile(configMapDir + "/" + configMap.getMetadata().getName() + "-" + namespace + ".log", configMap.toString());
