+++ systemtest/src/test/java/io/strimzi/systemtest/LogSettingST.java
@@ -28,6 +28,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import static io.strimzi.test.extensions.StrimziExtension.REGRESSION;
+import static io.strimzi.test.k8s.BaseCmdKubeClient.STATEFUL_SET;
 import static io.strimzi.systemtest.Constants.REGRESSION;
 import static io.strimzi.test.k8s.BaseKubeClient.STATEFUL_SET;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -173,11 +175,11 @@ class LogSettingST extends AbstractST {
         String eoName = entityOperatorDeploymentName(CLUSTER_NAME);
         String kafkaName = kafkaClusterName(CLUSTER_NAME);
         String zkName = zookeeperClusterName(CLUSTER_NAME);
-        Map<String, String> connectPods = StUtils.depSnapshot(NAMESPACE, connectName);
-        Map<String, String> mmPods = StUtils.depSnapshot(NAMESPACE, mmName);
-        Map<String, String> eoPods = StUtils.depSnapshot(NAMESPACE, eoName);
-        Map<String, String> kafkaPods = StUtils.ssSnapshot(NAMESPACE, kafkaName);
-        Map<String, String> zkPods = StUtils.ssSnapshot(NAMESPACE, zkName);
+        Map<String, String> connectPods = StUtils.depSnapshot(connectName);
+        Map<String, String> mmPods = StUtils.depSnapshot(mmName);
+        Map<String, String> eoPods = StUtils.depSnapshot(eoName);
+        Map<String, String> kafkaPods = StUtils.ssSnapshot(kafkaName);
+        Map<String, String> zkPods = StUtils.ssSnapshot(zkName);
 
         JvmOptions jvmOptions = new JvmOptions();
         jvmOptions.setGcLoggingEnabled(false);
