+++ systemtest/src/test/java/io/strimzi/systemtest/RecoveryST.java
@@ -16,6 +16,9 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
+import static io.strimzi.test.extensions.StrimziExtension.REGRESSION;
+import static io.strimzi.test.k8s.BaseCmdKubeClient.CM;
+import static io.strimzi.test.k8s.BaseCmdKubeClient.SERVICE;
 import static io.strimzi.systemtest.Constants.REGRESSION;
 import static io.strimzi.test.k8s.BaseKubeClient.CM;
 import static io.strimzi.test.k8s.BaseKubeClient.SERVICE;
