+++ test/src/main/java/io/strimzi/test/TestUtils.java
@@ -472,5 +472,4 @@ public final class TestUtils {
             }
         }
     }
-
 }
