+++ test/src/main/java/io/strimzi/test/k8s/KubeCluster.java
@@ -30,9 +30,9 @@ public interface KubeCluster {
     void clusterDown();
 
     /** Return a default CMD cmdClient for this kind of cluster. */
-    KubeClient defaultCmdClient();
+    KubeCmdClient defaultCmdClient();
 
-    Kubernetes defaultClient();
+    KubeClient defaultClient();
 
     /**
      * Returns the cluster named by the TEST_CLUSTER environment variable, if set, otherwise finds a cluster that's
