+++ test/src/main/java/io/strimzi/test/k8s/KubeClusterResource.java
@@ -21,8 +21,8 @@ public class KubeClusterResource {
 
     private final boolean bootstrap;
     private KubeCluster cluster;
-    private KubeClient cmdClient;
-    private Kubernetes client;
+    private KubeCmdClient cmdClient;
+    private KubeClient client;
     private HelmClient helmClient;
     private static KubeClusterResource kubeClusterResource;
 
@@ -42,10 +42,8 @@ public class KubeClusterResource {
     public static KubeClusterResource getKubeClusterResource() {
         if (kubeClusterResource == null) {
             kubeClusterResource = new KubeClusterResource();
-            return kubeClusterResource;
-        } else {
-            return kubeClusterResource;
         }
+        return kubeClusterResource;
     }
 
     /** Gets the namespace in use */
@@ -53,14 +51,14 @@ public class KubeClusterResource {
         return cmdClient().defaultNamespace();
     }
 
-    public KubeClient cmdClient() {
+    public KubeCmdClient cmdClient() {
         if (cmdClient == null) {
             this.cmdClient = cluster().defaultCmdClient();
         }
         return cmdClient;
     }
 
-    public Kubernetes client() {
+    public KubeClient client() {
         if (client == null) {
             this.client = cluster().defaultClient();
         }
