+++ test/src/main/java/io/strimzi/test/k8s/OpenShift.java
@@ -61,12 +61,12 @@ public class OpenShift implements KubeCluster {
     }
 
     @Override
-    public KubeClient defaultCmdClient() {
+    public KubeCmdClient defaultCmdClient() {
         return new Oc();
     }
 
-    public Kubernetes defaultClient() {
-        return new Kubernetes(new DefaultOpenShiftClient(CONFIG), "myproject");
+    public KubeClient defaultClient() {
+        return new KubeClient(new DefaultOpenShiftClient(CONFIG), "myproject");
     }
 
     public String toString() {
